Implementation of configuration synchronization manager
- Fixing night vision being turned on when the player hasn't bought it when using the binding.
- Fixing scavenger crashing during healing due to sound play.
